requirements.txt
anyio==4.3.0
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2024.2.2
click==8.1.7
dlib==19.24.4
exceptiongroup==1.2.1
face-recognition==1.3.0
face_recognition_models==0.3.0
ffmpeg==1.4
ffprobe==0.5
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==1.24.4
opencv-python==4.9.0.80
pillow==10.3.0
pydub==0.25.1
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2024.1
six==1.16.0
sniffio==1.3.1
tornado==6.4
typing_extensions==4.11.0
tzlocal==5.2
